fix more coderabbit review
