+++ packages/app-builder/src/components/ListAndTopicConfiguration/dataset-selection-provider-utils.ts
@@ -1,4 +1,7 @@
-import { ListAndTopicDatasetConfiguration } from './context/ListAndTopicDatasetConfiguration';
+import {
+  ListAndTopicDatasetConfiguration,
+  ListAndTopicDatasetConfigurationMode,
+} from './context/ListAndTopicDatasetConfiguration';
 
 export function makeDatasetsMap(selected: string[]): Record<string, boolean> {
   return Object.fromEntries(selected.map((name) => [name, true]));
@@ -6,7 +9,7 @@ export function makeDatasetsMap(selected: string[]): Record<string, boolean> {
 
 export function useListAndTopicDatasetConfigurationSharp(params: {
   datasets: Record<string, boolean>;
-  mode: 'view' | 'edit' | 'create';
+  mode: ListAndTopicDatasetConfigurationMode;
 }) {
   return ListAndTopicDatasetConfiguration.createSharp({
     datasets: params.datasets,
